Expose library system fields for library (#10)
Makes the library system name and a system-specific identifier for a
library or a service point available as part of library information and
query parameters. This is used for combining data from library directory
and other APIs for statistical purposes. At first this will be done with
Axiell's API.

For the fields and functionality for storing this information in Library
Directory see: libraries-fi/kirkanta#36